Working on creating a Profile Wizard, I created a 'suggestions' method.
ProfileUnits.suggestions aggregates suggestions from sublcasses.

Each subclass has a suggestions method that provides a list of suggestions.  These suggestions currently include a 'msg' attribute and a 'priority'.  The message is for display, the priority can be used to suggest the most important suggestions first.  Best guess values have been set of priority of suggestions, but may want to revisit in future.

 -- All tests pass.